+++ src/Ext/AnimType/Body.cpp
@@ -31,6 +31,7 @@ void AnimTypeExt::ExtData::LoadFromINIFile(CCINIClass* pINI)
 	this->CreateUnit_ConsiderPathfinding.Read(exINI, pID, "CreateUnit.ConsiderPathfinding");
 	this->XDrawOffset.Read(exINI, pID, "XDrawOffset");
 	this->HideIfNoOre_Threshold.Read(exINI, pID, "HideIfNoOre.Threshold");
+	this->Layer_UseObjectLayer.Read(exINI, pID, "Layer.UseObjectLayer");
 }
 
 const void AnimTypeExt::ProcessDestroyAnims(UnitClass* pThis, TechnoClass* pKiller)
@@ -106,6 +107,7 @@ void AnimTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->CreateUnit_ConsiderPathfinding)
 		.Process(this->XDrawOffset)
 		.Process(this->HideIfNoOre_Threshold)
+		.Process(this->Layer_UseObjectLayer)
 		;
 }