+++ src/Ext/AnimType/Body.h
@@ -26,6 +26,7 @@ public:
 		Valueable<bool> CreateUnit_ConsiderPathfinding;
 		Valueable<int> XDrawOffset;
 		Valueable<int> HideIfNoOre_Threshold;
+		Nullable<bool> Layer_UseObjectLayer;
 
 		ExtData(AnimTypeClass* OwnerObject) : Extension<AnimTypeClass>(OwnerObject)
 			, Palette(CustomPalette::PaletteMode::Temperate)
@@ -39,6 +40,7 @@ public:
 			, CreateUnit_ConsiderPathfinding(false)
 			, XDrawOffset(0)
 			, HideIfNoOre_Threshold(0)
+			, Layer_UseObjectLayer()
 		{ }
 
 		virtual ~ExtData() = default;