+++ docs/Whats-New.md
@@ -243,6 +243,8 @@ New:
 - Limbo Delivery of buildings (by Morton)
 - Ore stage threshold for `HideIfNoOre` (by Otamaa)
 - Image reading in art rules for all TechnoTypes (by Morton)
+- Attached animation layer customization (by Starkku)
+- Jumpjet unit layer deviation customization (by Starkku)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)